Moved partilce generating methods into own file
